# [MCC-488586] Optimize accessible ancestor objects