--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
+## 2.1.2
+* Optimized ActiveRecord adapter for PolicyMachine `#accessible_ancestor_objects`.
+
 ## 2.1.1
-* Added `include_prohibitions` option to PolicyMachine #scoped_privileges.
+* Added `include_prohibitions` option to PolicyMachine `#scoped_privileges`.
 
 ## 2.1.0
 * Elevate associations_filtered_by_operation to a public method in the ActiveRecord storage adapter.