--- a/lib/policy_machine_storage_adapters/active_record/postgresql.rb
+++ b/lib/policy_machine_storage_adapters/active_record/postgresql.rb
@@ -23,7 +23,7 @@ module PolicyMachineStorageAdapter
                 ,a.parent_id
               FROM assignments a
               JOIN ancestors anc
-      ON anc.parent_id = a.child_id
+                ON anc.parent_id = a.child_id
             )
             ),