Hi, this PR adds optimizations to the ActiveRecord Adapter's `accessible_ancestor_objects` method. It moves a lot of the logic from Ruby to raw SQL. In benchmarks with trivial amounts of data, this new method is approximately 3.5x as fast as the old method: 
```ruby
old = lambda { ado_pm.accessible_ancestor_objects(u1, bluff, grandparent_fish, key: :unique_identifier) }
new = lambda { ado_pm.policy_machine_storage_adapter.new_accessible_ancestor_objects(u1, bluff, grandparent_fish, key: :unique_identifier) }
Benchmark.ips do |x|
  x.report("old") { old.call }
  x.report("new") { new.call }
  x.compare!
end
<<-RESULTS
Warming up --------------------------------------
                 old     7.000  i/100ms
                 new    26.000  i/100ms
Calculating -------------------------------------
                 old     74.303  (± 2.7%) i/s -    378.000  in   5.092408s
                 new    264.500  (± 3.0%) i/s -      1.326k in   5.017673s

Comparison:
                 new:      264.5 i/s
                 old:       74.3 i/s - 3.56x  slower
RESULTS
```

@mdsol/team04 - please review