@hstebbins-mdsol 2019-04-22
dnm until i get performance results with more highly connected data