@DMcKinnon-mdsol 2019-04-22
> dnm until i get performance results with more highly connected data

Please also open a PR against dalton using this branch so we can see those specs pass too before we merge.