@hstebbins-mdsol 2019-04-23
@fpcyan was right! 😅 the existing implementation was quite slow against big data.

I've further modified the method to improve performance against large sets of data.

I've updated the OP, but now it's roughly 3.5x fast against trivial data. That being said, it's now faster against large amounts of data. These numbers are from a heavily assigned user (King Jim), getting scoped `read` privileges across 10 randomly selected CDSs he's assigned within. Note that this includes both direct assignments to CDSs and assignments to objects within a CDS.

```ruby
{"<cds uri>"=>{"ancestor_count"=>25, "result"=>{"old"=>1.53, "new"=>2.35}},
 "<cds uri>"=>{"ancestor_count"=>81, "result"=>{"old"=>1.82, "new"=>2.48}},
 "<cds uri>"=>{"ancestor_count"=>2092, "result"=>{"old"=>1.55, "new"=>1.67}},
 "<cds uri>"=>{"ancestor_count"=>63, "result"=>{"old"=>1.93, "new"=>2.52}},
 "<cds uri>"=>{"ancestor_count"=>1445},
 "<cds uri>"=>{"ancestor_count"=>76, "result"=>{"old"=>1.89, "new"=>2.45}},
 "<cds uri>"=>{"ancestor_count"=>311, "result"=>{"old"=>1.88, "new"=>2.48}},
 "<cds uri>"=>{"ancestor_count"=>1473, "result"=>{"old"=>1.34, "new"=>1.76}},
 "<cds uri>"=>{"ancestor_count"=>127, "result"=>{"old"=>1.84, "new"=>2.45}}}
```
Sorry for the raw data dump. Also, note that `result` values are for Iterations Per Second, _not_ elapsed time. So, a higher number == better performance.

The performance across the board is between 1.08x to 1.53x better with this new method, but 7/10 of these are around 1.30x better.

Note that one CDS is missing values: this is because attempting to execute the old method against that CDS results in a timeout (I'm testing against the read replica, which is much more sensitive to timeouts). That being said, the new method returns in 523ms.

Sorry for the delay in getting back to this, but I wanted to make sure I was getting good performance data. I now feel pretty confident about these performance improvements across the board in terms of data size.